Move middleware to root
