+++ cmd/gateway/data.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 	"distribuidos/tp1/protocol"
-	"distribuidos/tp1/server/middleware"
+	"distribuidos/tp1/middleware"
 	"distribuidos/tp1/utils"
 	"encoding/csv"
 	"errors"
