+++ cmd/gateway/result.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 	"distribuidos/tp1/protocol"
-	"distribuidos/tp1/server/middleware"
+	"distribuidos/tp1/middleware"
 	"fmt"
 )
