+++ middleware/messages_test.go
@@ -1,7 +1,7 @@
 package middleware_test
 
 import (
-	"distribuidos/tp1/server/middleware"
+	"distribuidos/tp1/middleware"
 	"reflect"
 	"testing"
 )
