+++ middleware/node/node.go
@@ -2,7 +2,7 @@ package node
 
 import (
 	"context"
-	"distribuidos/tp1/server/middleware"
+	"distribuidos/tp1/middleware"
 	"errors"
 	"fmt"
