+++ server/aggregators/groupByGame/main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"context"
-	"distribuidos/tp1/server/middleware"
-	"distribuidos/tp1/server/middleware/aggregator"
+	"distribuidos/tp1/middleware"
+	"distribuidos/tp1/middleware/aggregator"
 	"distribuidos/tp1/utils"
 	"maps"
 	"os/signal"
