+++ server/joiners/topNHistoricAvgJoiner/main.go
@@ -4,8 +4,8 @@ import (
 	"container/heap"
 	"context"
 	"distribuidos/tp1/protocol"
-	"distribuidos/tp1/server/middleware"
-	"distribuidos/tp1/server/middleware/joiner"
+	"distribuidos/tp1/middleware"
+	"distribuidos/tp1/middleware/joiner"
 	"distribuidos/tp1/utils"
 	"encoding/gob"
 	"os/signal"
